Delete methods for users, matches and messages

+++ grails-app/services/fr/mbds/firstgrails/UserCustomService.groovy
@@ -1,17 +1,24 @@
 package fr.mbds.firstgrails
 
+import fr.mbds.firstgrails.exceptions.UserNotFoundException
 import grails.gorm.transactions.Transactional
+import org.springframework.boot.context.config.ResourceNotFoundException
 
 @Transactional
 class UserCustomService {
+
     def userService
 
     def list(params) {
         return User.where { isDeleted == false }.list(params)
     }
 
-    def delete(Serializable id) {
+    def delete(Serializable id) throws UserNotFoundException {
         def user = userService.get(id)
+
+        if(!user)
+            throw new UserNotFoundException()
+
         user.isDeleted = true
         user.save(flush: true)
         UserRole.where { user == user }.deleteAll()
@@ -24,5 +31,10 @@ class UserCustomService {
     def get(Serializable id) {
         def user = User.get(id)
         return user ? (user.isDeleted ? null : user) : null
+    }
+
+    def userNotFoundException(final ResourceNotFoundException eexception) {
+        //logException exception
+        render view: 'notFound',  model: [id: params.id, exception: exception]
     }
 }
